+++ src/rules/anchor-has-content.ts
@@ -28,7 +28,7 @@ const rule: Rule.RuleModule = {
           },
           accessibleChildren: {
             type: "array",
-            items: { type: "string" },
+            items: { type: "string" }
           },
           accessibleDirectives: {
             type: "array",
@@ -41,8 +41,11 @@ const rule: Rule.RuleModule = {
   create(context) {
     return defineTemplateBodyVisitor(context, {
       VElement(node) {
-        const { components = [], accessibleChildren = [], accessibleDirectives = [] } =
-          context.options[0] || {};
+        const {
+          components = [],
+          accessibleChildren = [],
+          accessibleDirectives = []
+        } = context.options[0] || {};
 
         const elementTypes = ["a"].concat(components.map(makeKebabCase));
         const accessibleChildTypes = accessibleChildren.map(makeKebabCase);