+++ src/utils/hasContent.ts
@@ -7,7 +7,8 @@ import isHiddenFromScreenReader from "./isHiddenFromScreenReader";
 
 function hasDirective(node: AST.VElement, name: string) {
   return node.startTag.attributes.some(
-    (attribute) => attribute.directive && attribute.key.name.name === name.toLowerCase()
+    (attribute) =>
+      attribute.directive && attribute.key.name.name === name.toLowerCase()
   );
 }
 
@@ -26,13 +27,20 @@ function hasChildImageWithAlt(node: AST.VElement): boolean {
   });
 }
 
-function hasAccessibleDirective(node: AST.VElement, accessibleDirectives: string[]): boolean {
+function hasAccessibleDirective(
+  node: AST.VElement,
+  accessibleDirectives: string[]
+): boolean {
   return accessibleDirectives.some((directive) => {
-    return hasDirective(node, directive)
+    return hasDirective(node, directive);
   });
 }
 
-function hasContent(node: AST.VElement, accessibleChildTypes: string[], accessibleDirectives: string[]) {
+function hasContent(
+  node: AST.VElement,
+  accessibleChildTypes: string[],
+  accessibleDirectives: string[]
+) {
   return (
     hasAccessibleChild(node, accessibleChildTypes) ||
     hasAccessibleDirective(node, accessibleDirectives) ||